requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
Werkzeug==1.0.1
pymongo
opencv-python==4.1.2.30
Pillow==6.2.1
numpy
pandas
pytube
mysql-connector==2.2.9
pandas
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.31
Pygments==2.13.0
pytz==2022.2.1
urllib3
lxml
MarkupSafe
mysql-connector
click